requirements.txt
spotipy
python-dotenv
pandas
pyyaml
